+++ src/main/java/chess/dao/Member.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.dao;
 
 public class Member {
     private final String id;
