+++ src/test/java/chess/dao/MemberDaoTest.java
@@ -2,7 +2,6 @@ package chess.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.Member;
 import java.sql.Connection;
 import java.util.List;
 import org.junit.jupiter.api.Test;
